requirements.txt
python-telegram-bot==13.13
logging